rename of zenoh-flow-ctl into zfctl, updates to the Runtime trait
