+++ Cargo.toml
@@ -17,7 +17,7 @@ members = [
   "zenoh-flow",
   "zenoh-flow-derive",
   "zenoh-flow-daemon",
-  "zenoh-flow-ctl",
+  "zfctl",
   "cargo-zenoh-flow",
 ]
