+++ zfctl/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "zenoh-flow-ctl"
+name = "zfctl"
 version = "0.2.0"
 repository = "https://github.com/eclipse-zenoh/zenoh-flow"
 homepage = "http://zenoh.io"
@@ -9,7 +9,7 @@ authors = ["kydos <angelo@icorsaro.net>",
 edition = "2018"
 license = " EPL-2.0 OR Apache-2.0"
 categories = ["network-programming"]
-description = "Zenoh-Flow: zenoh-based data-flow programming framework for computations that span from the cloud to the device."
+description = "Zenoh-Flow command line tool"
 readme = "README.md"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -34,7 +34,7 @@ serde_derive = "1.0.55"
 serde = { version = "1.0.55", features = ["derive"] }
 serde_yaml = "0.8.13"
 serde_json = "1.0.55"
-serde-aux = "2.1.1"
+serde-aux = "3.0.1"
 log = "0.4"
 env_logger = "0.9.0"
 git-version = "0.3.4"
@@ -51,16 +51,16 @@ section = "utils"
 priority = "optional"
 assets = [
     # binary
-    ["target/release/zenoh-flow-ctl", "/usr/bin/zenoh-flow-ctl", "755"],
+    ["target/release/zfctl", "/usr/bin/zfctl", "755"],
 ]
 
 # RPM package configuration
 
 [package.metadata.rpm]
-package = "zenoh-flow-ctl"
+package = "zfctl"
 
 [package.metadata.rpm.cargo]
 buildflags = ["--release"]
 
 [package.metadata.rpm.targets]
-zenoh-flow-ctl = { path = "/usr/bin/zenoh-flow-ctl" }
+zfctl = { path = "/usr/bin/zfctl" }
